Added stop search function

+++ autosub/checkSub.py
@@ -38,52 +38,55 @@ class checkSub():
         # Initiate a session to OpenSubtitles and log in if OpenSubtitles is choosen
         if autosub.OPENSUBTITLES and autosub.OPENSUBTITLESUSER and autosub.OPENSUBTITLESPASSWD:
             OpenSubtitlesLogin()
-
         Index = 0
         End = len(autosub.WANTEDQUEUE)
         # loop through the wanted list and try to find subs for the video's
         # because we remove a video from the list we cannot use the internal counter from a for loop
         # so we track the position in the list with the variable 'Index'
-        while Index < End:
-            Wanted = {}
-            Wanted = autosub.WANTEDQUEUE[Index]
-            if not Wanted:
-                Index += 1
-                continue
+        if not autosub.SEARCHSTOP:
+            while Index < End:
+                if autosub.SEARCHSTOP:
+                    log.info('checkSub: Search stopped by User')
+                    break
+                Wanted = {}
+                Wanted = autosub.WANTEDQUEUE[Index]
+                if not Wanted:
+                    Index += 1
+                    continue
 
-            log.info("checkSub: Searching downloadlink(s) for %s, for %s" % (Wanted['file'], Wanted['langs']))
-            # get all links above the minimal match score as input for downloadSub
-            SubsNL,SubsEN = autosub.getSubLinks.getSubLinks(Wanted)
+                log.info("checkSub: Searching downloadlink(s) for %s, for %s" % (Wanted['file'], Wanted['langs']))
+                # get all links above the minimal match score as input for downloadSub
+                SubsNL,SubsEN = autosub.getSubLinks.getSubLinks(Wanted)
 
-            if not SubsNL and not SubsEN:
-                log.debug("checkSub: No subs found for %s" % Wanted['file'])
-                Index += 1
-                continue
-            if SubsNL:
-                log.debug('checkSub: Dutch Subtitle(s) found trying to download the highest scored.')
-                if DownloadSub(Wanted,SubsNL):
-                    Wanted['langs'].remove(autosub.DUTCH)
-                    if not autosub.DOWNLOADENG and autosub.ENGLISH in Wanted['langs']:
+                if not SubsNL and not SubsEN:
+                    log.debug("checkSub: No subs found for %s" % Wanted['file'])
+                    Index += 1
+                    continue
+                if SubsNL:
+                    log.debug('checkSub: Dutch Subtitle(s) found trying to download the highest scored.')
+                    if DownloadSub(Wanted,SubsNL):
+                        Wanted['langs'].remove(autosub.DUTCH)
+                        if not autosub.DOWNLOADENG and autosub.ENGLISH in Wanted['langs']:
+                            Wanted['langs'].remove(autosub.ENGLISH)
+                            SubsEN =[]
+                        if autosub.ENGLISHSUBDELETE and os.path.exists(os.path.join(Wanted['folder'],Wanted['file'] + Wanted['ENext'])):
+                            try:
+                                os.unlink(os.path.join(Wanted['folder'],Wanted['file'] + Wanted['ENext']))
+                                log.info("checkSub: Removed English subtitle for : %s" % Wanted['file'])
+                            except Exception as error:
+                                log.error("checkSub: Error while trying to remove English subtitle message is:%s." % error)
+                if SubsEN:
+                    log.debug('checkSub: English Subtitle(s) found trying to download the highest scored.')
+                    if DownloadSub(Wanted,SubsEN):
                         Wanted['langs'].remove(autosub.ENGLISH)
-                        SubsEN =[]
-                    if autosub.ENGLISHSUBDELETE and os.path.exists(os.path.join(Wanted['folder'],Wanted['file'] + Wanted['ENext'])):
-                        try:
-                            os.unlink(os.path.join(Wanted['folder'],Wanted['file'] + Wanted['ENext']))
-                            log.info("checkSub: Removed English subtitle for : %s" % Wanted['file'])
-                        except Exception as error:
-                            log.error("checkSub: Error while trying to remove English subtitle message is:%s." % error)
-            if SubsEN:
-                log.debug('checkSub: English Subtitle(s) found trying to download the highest scored.')
-                if DownloadSub(Wanted,SubsEN):
-                    Wanted['langs'].remove(autosub.ENGLISH)
-                    time.sleep(0)
+                        time.sleep(0)
 
-            if len(Wanted['langs']) == 0:
-                del autosub.WANTEDQUEUE[Index]
-                time.sleep(0)
-                End -= 1
-            else:
-                Index += 1
+                if len(Wanted['langs']) == 0:
+                    del autosub.WANTEDQUEUE[Index]
+                    time.sleep(0)
+                    End -= 1
+                else:
+                    Index += 1
 
 
         autosub.DBCONNECTION.close()
@@ -98,4 +101,5 @@ class checkSub():
         log.info("checkSub: Finished round of subs Search. Go to sleep until the next round.")
         autosub.SEARCHTIME = time.time() - StartTime
         autosub.SEARCHBUSY = False
+        autosub.SEARCHSTOP = False
         return True
